+++ Editor/EditorCore/ProBuilderEditorShortcuts.cs
@@ -10,21 +10,30 @@ namespace UnityEditor.ProBuilder
 	/// </summary>
 	static class ProBuilderEditorShortcuts
 	{
-		[Shortcut("ProBuilder/Editor/Edit Vertices", typeof(PositionToolContext.ProBuilderShortcutContext))]
+		[Shortcut("ProBuilder/Editor/Edit Vertices", typeof(SceneViewMotion.SceneViewContext))]
 		static void SetSelectMode_Vertex()
 		{
+			if (!CheckAndEnterPBContextIfNeeded())
+				return;
+
 			ProBuilderEditor.selectMode = SelectMode.Vertex;
 		}
-
-		[Shortcut("ProBuilder/Editor/Edit Edges", typeof(PositionToolContext.ProBuilderShortcutContext))]
+		
+		[Shortcut("ProBuilder/Editor/Edit Edges", typeof(SceneViewMotion.SceneViewContext))]
 		static void SetSelectMode_Edge()
-		{
+		{			
+			if (!CheckAndEnterPBContextIfNeeded())
+				return;
+			
 			ProBuilderEditor.selectMode = SelectMode.Edge;
 		}
-
-		[Shortcut("ProBuilder/Editor/Edit Faces", typeof(PositionToolContext.ProBuilderShortcutContext))]
+		
+		[Shortcut("ProBuilder/Editor/Edit Faces", typeof(SceneViewMotion.SceneViewContext))]
 		static void SetSelectMode_Faces()
 		{
+			if (!CheckAndEnterPBContextIfNeeded())
+				return;
+			
 			ProBuilderEditor.selectMode = SelectMode.Face;
 		}
 
@@ -33,6 +42,20 @@ namespace UnityEditor.ProBuilder
 		{
 			if(ProBuilderEditor.instance != null)
 				ProBuilderEditor.instance.ToggleSelectionMode();
+		}
+
+		static bool CheckAndEnterPBContextIfNeeded()
+		{
+			if (ToolManager.activeContextType != typeof(PositionToolContext))
+			{
+				// Check if PositionToolContext can actually be entered
+				if (EditorToolManager.GetComponentContext(typeof(PositionToolContext), true) == null)
+					return false;
+				
+				ToolManager.SetActiveContext<PositionToolContext>();
+			}
+
+			return true;
 		}
 	}
 }